ASoC: msm8x10-wcd: Update codec driver
Update codec driver to remove virtual port widgets, add
device tree parsing functions and route table, and fix playback
and recording issue.

+++ sound/soc/codecs/msm8x10-wcd.h
@@ -199,7 +199,7 @@ struct msm8x10_wcd {
 	int (*read_dev)(struct msm8x10_wcd *msm8x10,
 			unsigned short reg, unsigned int *val);
 	int (*write_dev)(struct msm8x10_wcd *msm8x10,
-			 unsigned short reg, unsigned int val);
+			 unsigned short reg, u8 val);
 
 	u32 num_of_supplies;
 	struct regulator_bulk_data *supplies;
